requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1

# Database
asyncpg==0.29.0
redis==5.0.1

# AWS Dependencies
boto3==1.34.0
botocore==1.34.0


# Authentication
authlib==1.2.1
itsdangerous==2.1.2

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Testing
pytest-cov==4.1.0
pytest-mock==3.12.0 